move functions to pyscftools and replace mol/pbc functions
