+++ pyqmc/slater.py
@@ -2,6 +2,7 @@ import numpy as np
 import pyqmc.gpu as gpu
 import pyqmc.determinant_tools as determinant_tools
 import pyqmc.orbitals
+import pyqmc.pyscftools
 import warnings
 
 
@@ -127,7 +128,7 @@ class Slater:
             self._det_occup,
             self._det_map,
             self.orbitals,
-        ) = pyqmc.orbitals.choose_evaluator_from_pyscf(
+        ) = pyqmc.pyscftools.orbital_evaluator_from_pyscf(
             mol, mf, mc, twist=twist, determinants=determinants, tol=self.tol
         )
