+++ tests/unit/test_wf_derivatives.py
@@ -166,7 +166,7 @@ def test_manual_pbcs_correct(H_pbc_sto3g_kuks, epsilon=1e-5, nconf=10):
     This test makes sure that the number of k-points must match the number of k-points
     in the mf object.
     """
-    from pyqmc.determinant_tools import create_single_determinant
+    from pyqmc.pyscftools import single_determinant_from_mf
 
     mol, mf = H_pbc_sto3g_kuks
     supercell = np.identity(3, dtype=int)
@@ -174,8 +174,8 @@ def test_manual_pbcs_correct(H_pbc_sto3g_kuks, epsilon=1e-5, nconf=10):
     mol = pyq.get_supercell(mol, supercell)
 
     determinants = [
-        create_single_determinant(mf, 1.0)[0],
-        create_single_determinant(mf, -0.2)[0],
+        single_determinant_from_mf(mf, 1.0)[0],
+        single_determinant_from_mf(mf, -0.2)[0],
     ]
     for s, ka, a, ki, i in [(0, 0, 0, 0, 1)]:
         determinants[1][1][s][ka].remove(a)
